clean code, use same add for add cmd and build install
